magit-toplevel: fix gitdir of submodule over Tramp case
When the control directory of a submodule is located inside the
.modules/ subdirectory of the control direct of the super repository,
then `git rev-parse --show-cdup' returns an absolute filename, which
makes `expand-file-name' on that filename a no-op.

When that directory is on a remote host accessed using Tramp, then
that means that `expand-file-name' does not expand the filename to
also include the remote part of `default-directory'.

So we have to add that explicitly.

Fixes #2305.