magit-process-find-password-functions: new option
Add new option `magit-process-find-password-functions' and advertise
adding `magit-process-password-auth-sources' to its value, but do not
actually do so.

We don't use that function by default, because most users will be
unaware of this functionality and that might cause issues.  For example
a user may have configured `auth-source' at some point, and then stopped
using it and also changed the password without also changing it in the
source used by `auth-source'.  In such a case `auth-source' would return
the wrong password and that would then be feed to git.  The user likely
would not note that this is what is happening and report a Magit defect.