magit-branch-spinoff: Check if branch exists
Abort if a branch with the new name already exists.  Otherwise, creating
the new branch fails, but the current branch is moved back to the merge
base it shares with its upstream branch.  At this point, the reflog is
needed to restore commits between the starting point and the merge
base (unless another ref happens to be on the starting commit).