handle custom cygwin mounts
Rely on the output from mount to tell us how to map cygwin paths into
native Windows paths.

This is also needed in git-commit; avoid a hard dependency on magit by
only loading magit-git in the case where we receive a bad path to
foo_EDITMSG (and falling back to the standard "/cygdrive" prefix if we
fail to load).