magit-blame: save a modified file-visiting buffer
If a buffer exists which visits the file to be blamed and if it has
unsaved changes, then unconditionally save it.  Saving the buffer is
necessary because Git otherwise would not know about the modifications,
and so the line numbers it prints would not correspond to the line
numbers in the buffer.