magit-stash-drop: verbosely to be able to undo mistakes
When a branch is deleted then the old hash is printed to the process log
buffer.  We would like to do the same thing here, but unfortunately that
is not possible because `git reflog delete --verbose ...' does not print
the commit hash as one would expect.  Instead it prints the commit
message, which is quite useless.

So instead we have to print similar information using `message'.

Re #2383.