magit-revert-buffers: avoid unnecessary remote connections
Before checking whether a file is inside the repository for which
buffers are being refreshed, first check whether it is at least located
on the some host, using `file-remote-p'.  This optimization is necessary
because `file-in-directory-p' calls `file-truename' on both the file as
well as the directory, which, for remote files, requires a round trip to
the remote host(s).  (It would be better if `file-in-directory-p' itself
implemented that optimization, but for now we have to do it ourselves.)

Re http://emacs.stackexchange.com/questions/17929.