stop faking negative overlay property priorities
Face attributes set in overlays always have a higher priority than the
corresponding text properties.  Magit highlights the current section
with an overlay whose face sets the background color, as a result all
text in the current section would lose its background color.

We used to fake negative priorities for such text which has a background
color, by using an overlay to set the face where one would usually use a
text property.  So instead of giving the highlighting overlay a negative
priority, i.e. one lower than the 1 priority of text properties (because
that's impossible) we instead gave the text properties a higher priority
(by using overlay properties instead).  Conceptually that was equivalent
to giving the highlighting overlay a negative priority.

But we should not have done that for three reasons:

- Giving ref labels (and a few other things) a background color caused
  the buffer to be much noisier than it otherwise would have been.  So
  at least in my opinion, giving labels a background color was not only
  a hack, and difficult and expensive, it was also undesirable.  In the
  v2.1.0 release I dropped the background color of all ref labels and a
  few other faces.  This was generally well received.

- Using overlays is inefficient, and should therefore be avoided when
  not necessary.  Every time Magit comes up Eli tells us about it.

- Using overlays complicates the code and makes necessary changes more
  difficult.

Dropping support for faked negative overlay priorities, might mean that
some third-party themes have to be adjusted.  Which is why I did not do
it in v2.1.0, but now the time has come.