magit-define-popup: add :setup-function and :refresh-function
Deprecate the optional, positional argument MODE.  That was used in
the past to set the buffer-local value of `magit-popup-setup-hook'.
That hook is also deprecated now, but, the mode, still respected if set.

Instead add the new keyword arguments `:setup-function' (which replaces
`magit-popup-setup-hook', but is ineffective if that is non-nil) and
`:refresh-function' (for which no equivalent existed before).