stop displaying hashes in status headers
Do not display hashes in the "Head" and "Upstream" headers displayed in
the status buffer, except when HEAD is detached, in which case the hash
is displayed where usually the name of the current branch would be
displayed.

Displaying the hashes is distracting and also unnecessary (because one
can press C-w to echo the hash, and also put it on the `kill-ring' and
the `magit-revision-stack'.