magit-push-current: fallback to magit-push-elsewhere or magit-push
If HEAD is detached, then fall back to `magit-push'.  If the current
branch has no upstream, then fall back to `magit-push-elsewhere'.
That is a temporary approach, eventually this command will raise an
error if HEAD is detached or the upstream is not configured.

This was the last command which still used `magit-push-read-args',
remove that now.