improve transfer popups and their actions
Improve the popups `magit-fetch-popup', `magit-pull-popup',
`magit-push-popup' and `magit-rebase-popup', and their actions.

Each popup feature three actions which acts on the current branch and
the push-remote, the upstream, or any other source or target.  Some of
these popups also feature other actions and continue to do so.

If one of the mentioned actions isn't usable, because the branch it is
supposed to act on is not currently configured, then it is hidden.  If
such a command is invoked while it is unusable, then that now results
in a user-error.

The commands are renamed and the bindings are changed, both to increase
consistency among these popups and to be make them more self-explanatory.
Three of these commands were previously missing.

                                p: `magit-fetch-from-pushremote'
  f: `magit-fetch-current'   => u: `magit-fetch-from-upstream'
  o: `magit-fetch'           => e: `magit-fetch'

                                p: `magit-pull-from-pushremote'
  F: `magit-pull-current'    => u: `magit-pull-from-upstream'
  o: `magit-pull'            => e: `magit-pull'

  Q: `magit-push-quickly'    => p: `magit-push-current-to-pushremote'
  P: `magit-push-current'    => u: `magit-push-current-to-upstream'
  e: `magit-push-elsewhere'  => e: `magit-push-current'

                                p: `magit-rebase-onto-pushremote'
  l: `magit-rebase-unpushed' => u: `magit-rebase-onto-upstream'
  r: `magit-rebase'          => e: `magit-rebase'

The bindings of some other actions were also adjusted.

  `magit-rebase-subset':      o => s
  `magit-rebase-edit-commit': e => m