branch.autoSetupPush: new Git variable
Add new function `magit-maybe-set-branch*pushRemote' and call it from
the commands `magit-branch' and `magit-branch-and-checkout'.  When the
variable `branch.autoSetupPush' is set, then this function sets the
value of `branch.<name>.pushRemote' to the same value.  The value has
to be an existing branch, if it is not, then it is ignored.

`branch.autoSetupPush' is a Magit invention, Git knows nothing about it.
It was inspired by the standard Git variables `branch.autoSetupMerge'
and `branch.autoSetupRebase'.