magit-branch-popup: add the relevant variables
Add all variables to the popup that are relevant when creating a new
branch.  This allows users to see what the current values are and also
to change the values.

Also add the variables related to the current branch and the variables
that are consulted when the branch-specific variables are unset.

For each variable define the necessary command and format function.
These commands have to form `magit-VERB-VARIABLE'.  VERB is one of
"set", "edit", or "cycle".  VARIABLE is the name of the Git variable.
For branch-specific variables, VARIABLE is `branch*SUFFIX', where
"*" stands for ".<name>."

Remove `magit-branch-unset-upstream', and replace
`magit-branch-set-upstream' with `magit-cycle-branch*pushRemote' and
`magit-branch-edit-description' with `magit-edit-branch*description'.