remove magit-specific git-variable branch.autoSetupPush
It turned out that this is not actually necessary but has downsides.

Users should instead set `remote.pushDefault' once per repository.
Doing so has the advantage that this variable affects all branches,
including branches that already exist.  `branch.autoSetupPush' on the
other hand only takes effect when new branches are being created.

Furthermore needlessly setting a variable per-branch (using
`branch.<name>.pushRemote') which could just as well be set per
repository (because it is the same for all branches) has the
disadvantage that changing that value becomes much more work.

In rare cases the push-remote of a branch might be different from that
of the other branches.  In that case `branch.<unicorn>.pushRemote'
should be set manually.  Whether `branch.autoSetupPush' exists or not,
doesn't matter, if the unicorn needs a different push-remote, then
`branch.<unicorn>.pushRemote' always has to be set explicitly.