magit-popup-format-action-button: preserve existing description face
When the return value of a function is used as the description of
the action, then check if any part of the string already has a face
set and if so then just use the string as is.

Otherwise continue to assume that the value is a branch name and use
the appropriate face.  This fallback is a kludge, which assumes that
the description always is a branch.  That does not have to be the
case but at present there are no popups for which this assumption
does not hold true.