generalize visibility cache
Currently only the visibility of `file' sections is being cached.
Generalized the code used to do so to make it possible to use the
cache for other types too.

This is done by renaming the involved symbols and adding convenience
utilities.  Also `magit-section-show' now explicitly removes entries
from the cache.

Despite the names of the new functions, it's still only possible to
store the information that a section, which by default is expanded,
should be hidden when it reappears.  That might change eventually, so
that it becomes possible to remember that sections which are hidden by
default should be expanded when they reappear.  But we don't need that
currently.

Also note that this feature should only be used selectively, using it
for every section would not be good for performance.