magit--push-current-to-pushremote-desc: correct description
Previously the description contained "@{push}", which isn't precise.
What is being pushed by the respective command is configured using
either `branch.<name>.pushRemote' or `branch.pushDefault'.  So we
should use "pushRemote" as the shorthand.  If these variables are
set, then "@{push}" is usually based on their values, but when they
are unset, then "@{push}" might be defined anyway, based on other
variables.

"@{push}" simply means "where `git push' would push to" (provided a
single branch is being pushed), and not "the branch on the configured
push remote we would push to".