magit--push-current-to-pushremote-desc: correct description
Previously the description contained "@{push}", which is correct
but not precise.  What is being pushed is configured using either
`branch.<name>.pushRemote' or `branch.pushDefault'.  So we now use
"pushRemote" as the shorthand.  If these variables are set, then
"@{push}" is usually based on their values, but when they are unset,
then "@{push}" might be defined anyway, based on other variables.

"@{push}" simply means "where we would push to" (provided a single
branch is being pushed), and not "the branch on the configured push
remote we would push to".