better defaults for reference completion
Add the new functions `magit-name-branch', `magit-name-local-branch',
`magit-name-remote-branch' and `magit-name-tag', and use them in the
respective `magit-{*}-at-point' functions.

This affects the various `magit-read-{*}' completions functions which
now offer the branch or tag at point as the default completion candidate
even when the type of the section at point is `commit'.