magit-branch-prefer-remote-upstream: use saner interpretation
Do not consult `magit-branch-prefer-remote-upstream' when determining
the default choice when reading the starting-point on behalf of
`magit-branch' and `magit-branch-and-checkout'.  Just offer the branch
at point as default.  Remote branches are preferred, but no remote
branch is substituted for a local branch.

Instead make these commands behave similar to how `magit-branch-spinoff'
already behaves.  Do use the chosen starting-point as starting-point,
but possibly set the upstream of the starting-point as the upstream of
the new branch.