magit-pull-or-fetch-popup: new popup
Combine `magit-pull-popup' and `magit-fetch-popup' into one common
popup.  By default the old popups are still bound in `magit-mode-map'.

Add this to your init file to use the new popup:

  (with-eval-after-load 'magit-remote
    (define-key magit-mode-map "f" 'magit-pull-and-fetch-popup)
    (define-key magit-mode-map "F" nil))

One advantage of the new popup is that it is once more possible to
press the same key twice to fetch.  However that binding now is "f f",
not "F F", and it fetches all remotes not just the current remote.

The new popup also lacks the commands for fetching from only a single
remote.  You can add them to the new popup by adding something like this
to your init file:

  (with-eval-after-load 'magit-remote
    (magit-define-popup-action 'magit-pull-and-fetch-popup ?U
      'magit-get-upstream-branch
      'magit-fetch-from-upstream-remote ?F)
    (magit-define-popup-action 'magit-pull-and-fetch-popup ?P
      'magit-get-push-branch
      'magit-fetch-from-push-remote ?F))