magit-push-implicitly: restore command
This command was removed earlier, but it turned out that it is still
required for certain work-flows.  Restore the command, but do not add it
to the push popup again.  If the user follows the instructions on how to
add the command to the popup, then that now shows what the command will
do.