magit-revert-buffers-only-for-tracked-files: change default value
Go back to only reverting the tracked files.  I expect that for most
users reverting all files would be better (because that improves
performance), but because there is a tiny change of data loss with that
setting when combined with odd work flows I am changing the default.

Unfortunately that means that thousands of users will needlessly suffer
from bad performance, and they will certainly complain about that, but
that criticism will be pleasant in comparison to what I would have to
expect from users who lost data.