revert: no longer update the VC information in the mode-line
No longer add `magit-refresh-vc-mode-line' to the default value of
`magit-after-revert-hook' and `magit-not-reverted-hook', because
it causes a noticeable delay when there are many open buffers or
many tracked files.

I assume that most Magit users rely on the status buffer for status
information, not VC's mode-line information.  User who rely on that
information will have to customize the hooks.