require magit in log and diff commands
The dependencies between magit libraries are not strictly linear, which
is problematic when someone uses a command that is defined in a library
that is a dependency of `magit', but that also depends on code located
in `magit', relying on autoload definitions and without first using a
command which causes `magit' to be loaded.

To work around that, require `magit' in the body of such commands.
Actually, because the diff and log entry-point commands all use
`magit-diff-setup' or `magit-log', `magit' is required inside those
functions.