magit-update-ref: new function
By default Git only maintains reflogs for refs in certain namespaces.
To force it do so for other refs it was necessary to manually create the
appropriate file .git/logs/<ref> before calling `git update-ref'.  Since
v2.6.0 that command supports the `--create-reflog' argument, and we now
use it if available.

To make that easier, wrap `git update-ref' in the new function
`magit-update-ref', which also falls back to the old approach for older
Git versions.  Also remove the function `magit-reflog-enable', which is
where the hack used to be located.

This also fixes #2515.  The old hack failed when creating a stash in a
linked working tree when the repository did not already contain another
stash.  The problem was that a file .git/worktrees/<worktree>/logs/<ref>
was created, instead of .git/logs/<ref>.  Like the `--create-reflog'
argument, support for linked working trees was added in Git v2.6.0, so
it's okay to use the old "touch file" approach for older Git versions
(not that we have a choice anyway).