magit-branch-reset: do not set upstream by default
Resetting a branch to another branch, causes the former to be configured
as the upstream of the latter.  That behavior was introduced by accident
when the argument `--track' was replaced with the variable
`branch.autoSetupMerge' in the popup.  Because it would be undesirable
to restore the `--track' argument in the popup, one now has to use a
prefix argument to set the upstream while resetting.

Re #2517.