magit-clone: no longer create symbolic-ref origin/HEAD by default
Actually, since `git clone' always creates that symbolic-ref,
automatically *remove* that symbolic-ref right after cloning.

The default is to remove that symbolic-ref because it isn't actually
very useful.  Right after running `git clone' one knows which of the
remote branches is the "remote HEAD" without having to rely on that
symbolic-ref, because the local branch is automatically setup to use
that as upstream.  At all other times that symbolic-ref is unreliable
because it is not automatically updated then the "remote HEAD" changes.

But some users might want it anyway, so add a new option
`magit-clone-set-remote-head', which can be set to `t' to preserve the
symbolic-ref.