magit-maybe-show-margin: ensure the margin is restored
If the buffer-local value of `magit-show-margin' is set, then call
`magit-set-buffer-margin' with its value as ENABLE value.

In [1: f77325e] I thought it would be enough to just forgo calling
`magit-set-buffer-margin' again with the default value of
`magit-{log,reflog}-show-margin', but that is only the case if the
buffer is still being displayed in the same window as before.  But
since margins are local to windows, we actually have to recreate
them when we might just have displayed the buffer in a new window.

Note that this only affects the selected window.  All other windows
which might be displaying the same buffer are not affected, but I've
decided that this is a feature.

1: f77325ea13e68afce79801243fa3f8a4a463bc7f
   magit-maybe-show-margin: do nothing if magit-show-margin is set