magit-cygwin-mount-points: check git --exec-path
Don't just rely on the existence of 'mount' to determine whether we need
to handle cygwin style paths.  First, check that 'git --exec-path'
returns a non-Windows path, which is a much better indicator.  This way
we correctly handle the case where Git for Windows is used,
and (non-git) cygwin utilities are installed on the PATH.

Fixes #2534.