magit-get-next-tag: never log output
When `magit-git-debug' was non-nil, then this function always logged
that `HEAD' cannot be described and presented that in a way that
suggests that this is an error.  It is not; this command only exits with
a zero exit status when the `HEAD' is tagged or reachable from a tag, so
it is expected that this command "fails" most of the time.