magit-auto-revert-mode: add kludge to avoid turning it on
The global minor-mode `magit-auto-revert-mode' is turned on by default,
which is something one usually shouldn't do for global minor-modes.
There already where some necessary kludges in place to make this work.

Now, if the user sets the variable `magit-auto-revert-mode' or the
obsolete alias `magit-revert-buffers' to nil anytime during startup,
before or after `magit-autorevert' is loaded, then the mode is not
turned on (actually it is explicitly turned off).  Previously there was
a similar kludge in use, which did this only for `magit-revert-buffers'.

Re #2559.