magit-diff-wash-submodule: rewrite from scratch
The old implementation was a hack, which got extended when I got myself
into a situation where the current version failed.  This time around it
was "(not checked out)" which was not handled properly.  Upon inspecting
`show_submodule_summary' in `submodules.c' it turned out that "(revision
walker failed)" also wasn't handled, and why there sometimes is a
trailing colon and sometimes not (there's a colon when commits follow,
so the colon cases should be handled separately even though that leads
to a bit of duplication).

At present, when `diff.submodule' is `log', then these are the lines we
have to look for.

    Submodule <module> contains untracked content
    Submodule <module> contains modified content
    Submodule <module> <from>...<to>:
    Submodule <module> <from>..<to>:
    Submodule <module> <from>..<to> (rewind):
    Submodule <module> <from>..<to> (submodule deleted)
    Submodule <module> <from>..<to> (not checked out)
    Submodule <module> <from>..<to> (new submodule)
    Submodule <module> <from>..<to> (commits not present)
    Submodule <module> <from>..<to> (revision walker failed)

This commit also changes how some of these variants are represented in
Magit.  Previously some information was lost in some cases.

When `diff.submodule' is `short', then there are no lines matching
"^Submodule" and `magit-diff-wash-diff' takes care of washing it.