magit-insert-stash-untracked: recurse into sub-trees
Without the -r flag, ls-tree will return the top-level directory name
rather than the file name for an untracked file that is not in the
top-level directory.  As a result, the diff of stash@{N}^3 is limited to
the directory rather than the untracked file.  This leads to all other
files in directory being shown as deleted.