magit-diff: hide trailing CR chars on Windows
Windows users are more likely to interact with CRLF files, in which case
their diffs will have trailing CR characters. We keep them around, as
they are useful for staging (or otherwise applying) hunks without
mangling line endings but users are usually not interested in seeing a
diff full of trailing ^Ms, so we make them invisible.

This behavior can be disabled using the `magit-diff-hide-cr-characters'
variable.