magit-read-{remote,upstream}-branch: remove duplicate choices
In both cases, the collection to choose from is created from an
argument (local branch) turned into a remote ref, plus a list of all
remote or all branches.  The latter part can already contain the refs
constructed in the former, leading to duplicates.

Since we use dash anyway and the order in the collection is irrelevant,
replacing nconc by -union should get rid of these duplicates.