add ISSUE_TEMPLATE file
This file is inserted into the text area when a user opens a new issue,
which considerably increases the odds that these instructions are
actually being read.

Previously, when Github did not support this new file yet, we displayed
similar, but less complete, information at the top of the README.md in
the hope that users would see it there, which wasn't always the case.

Remove the information which is now part of ISSUE_TEMPLATE from
README.md and CONTRIBUTING.md.