stash: clarify worktree-related descriptions
magit-stash-worktree, magit-snapshot-worktree, and
magit-insert-stash-worktree are limited to unstaged changes in the
stash (stash@{N}^2..stash@{N}).  Update their doc-strings to refer to
these changes as "unstaged".  Likewise, rename the section "Working
tree" in magit-stash-mode to "Unstaged".

For consistency, describe the index counterparts as "staged".

Although it would be clearer to also rename the commands as "*-unstaged"
and "*-staged" rather than "*-worktree" and "*-index", the keys in the
popup are aligned with the current names (w, W -> worktree, i, I ->
index).  Stick with the current names to avoid either 1) extensive
changing of the popup keys or 2) having keys that don't have an obvious
connection to the command name.