magit-diff-wash-submodule: use appropriate range for rewinds
When `diff.submodule' is `log', then `git diff' prints one of

    Submodule <module> <from>...<to>:
    Submodule <module> <from>..<to>:

for modules with new commits, as appropriate.  But for rewinds it
always prints:

    Submodule <module> <from>..<to> (rewind):

In that case `git log --online --left-right <from>..<to>' prints
nothing.  Inject the additional period to get the same range and
similar output as `git submodule summary'.

* <path> <from>...<to> (<n>):
  < removed commit