git-commit-propertize-diff: use font-lock-ensure
36be43d (use font-lock-flush instead of font-lock-ensure, 2015-08-15)
changed this to use font-lock-flush.  However, based on tests with a
somewhat recent build of the emacs-25 branch (cf79616) and on this
discussion [1], I don't think font-lock-flush will reliably highlight
the diff.  Also, the initial issue that 36be43d was addressing is, I
think, due to a bug in Emacs's font-lock-ensure (bug#22399) that has now
been fixed (21beb19, 2016-01-25).

[1]: https://lists.gnu.org/archive/html/emacs-devel/2016-02/msg01481.html