magit-merge-nocommit: always use --no-ff
Use `--no-ff' even if it is not specified in the popup.  Otherwise Git
would to a fast-forward merge if possible, which clearly isn't what one
would expect to happen.  The intention is to create a merge commit, but
have the opportunity to inspect the merge before creating that commit.
To not create the commit but still perform (and finish) the merge, is
something very different.

Fixes #2620.