magit-diff-wash-diff: fix -L kludge false positive
f06edb2 (magit-log-{popup,buffer-file}: add ‘-L’ argument, 2015-09-30)
stripped the first two characters of the file name to account for
--no-prefix not working with when -L is used with git log.  However,
magit-diff-wash-diff is also called in log mode when --patch is used
without --graph, and in this case --no-prefix has an effect, so the file
name gets truncated.  Check that -L is in magit-refresh-args to avoid
this.

Fixes #2622.