magit-commit-squash-internal: commit synchronously before rebase
Previously it was possible for the commit process to still hold the lock
on index.lock, when the rebase process was trying to acquire it.  Now we
run the commit process synchronously if we are going to also rebase, and
we also forgo refreshing between the two processes.  If we don't rebase,
then still commit asynchronously and refresh.