magit-submodule-add: offer relative paths for completion
Instead of using `read-file-name', which allows selecting a path which
is not a repository, use `magit-completing-read' to select the relative
name of a known repository located inside the outer repository.

This also fixes a second issue: Previously an absolute path may have
been used as initial input.  If the user used that and the chosen
repository was not located at the top-level of the outer repository,
then `read-file-name' considered that an invalid choice and prompted
again, this time without a default.