use --reject to apply at least non-conflicting hunks
When applying changes to files in the working tree results in conflicts,
then `git apply' aborts and does not apply anything.  What we would like
it to do, is to perform a 3-way merge, just like when applying changes
to the index, but unfortunately that is not possible.

When reversing changes and when discarding staged changes while there
are unstaged changes, then use `--reject'.  This won't insert conflict
markers into files in the working tree, but it ensures that changes that
can be applied are applied even when there are other changes that cannot.

For changes that could not be applied `--reject' causes the creation of
diff files which contain those changes.  We would have liked for Git to
create conflict markers - using such "conflict files" is the closest we
can get to that.