magit-commit-diff: make sure diff buffer doesn't replace message buffer
When the window that just began displaying the commit buffer previously
displayed the diff buffer, then displaying that buffer again would
likely be done by displayed in that same window, causing the commit
message buffer would no longer be displayed anywhere.  Prevent that by
removing the diff buffer from the window's list of previously displayed
buffers.

Also remove the kludge in `magit-display-buffer-traditionally', which
was also supposed to deal with this issue.