magit-buffer-locked-p: use boolean as value
Always set `magit-buffer-locked-p' to a boolean value instead of storing
the locked value or nil.  Calculating that value is cheap, so we don't
have to cache it.  And if we did cache it, then we would have to update
the cache when the value changes (which we failed to do).