magit-popup-describe-function: use correct window
If we let `display-buffer' (called by `describe-function') decide where
to put the *Help* window, it may (if the frame is wide enough) decide to
split horizontally instead of using the window we were creating earlier
with `split-window-below'.  This results in 3 windows, 2 of which are
the popup buffer, and is generally not very helpful.

To avoid this, let-bind `display-buffer-alist' to tell `display-buffer'
that it should reuse the window we made.