add more magit-display-buffer-function options
As mentioned in #2541, there are a lot of possible behaviors for
display-buffer functions.  Provide a few more options that users can try
before resorting to writing their own functions.

magit-display-buffer-same-window-except-diff-v1 and
magit-display-buffer-fullcolumn-most-v1 are modified from variants
proposed in #2541.  magit-display-buffer-fullframe-status-v1 provides
the full-frame status behavior that seems to be popular (#1953).

The names of these functions are versioned, as suggested in #2541.

The helper function magit--display-buffer-fullframe is modified from a
function that Sebastian Wiesner shared in #1953.