magit-branch-read-args: don't default to existing
Setting magit-branch-read-upstream-first to non-nil is useful when
creating a local branch of an existing remote topic branch, because it
gives the remote name as the initial-input.  When creating a new topic
branch from origin/master this suggests "master" which is invalid since
that branch (probably) already exists.

Therefore, avoid giving an initial-input when that would be an existing
branch.